GCM notifications, turn camera on/off
Also now sending data to meteor instead of netcat.
Starting up the app, then pressing the back button, and repeating cuases
the app to crash because of out of memory. The reason is that
CameraPreview is allocating some buffers and never freeing them. I'm
not sure why it's not freeing it through the garbage collector - there
is something that continues to hold a reference to the data.

But when using camera_off and camera_on commands, we just hack
in `process.exit`, which kills the app enough that the memory does get
freed.